Don't delete the log files and make a nice tar.bz2 archive for downloading.
Apparently the lookat script symlinks to the local copy of the log file
when it unpacks, instead of symlinking to the actual log file. Thus, you
can't delete the log file symlinks that are retreived by get-submissions.

Also, this now archives the submissions in a format that is easily
downloaded onto a home computer.